fix: -Wsign-compare in the polyline/circle geometry, 11 -> 0
Polyline.cpp 5 -> 0, Geometry/Circle.cpp 4 -> 0, MultiPoint.cpp 1 -> 0,
SVG.cpp 1 -> 0. Circle.cpp and SVG.cpp now have no GCC warnings of any kind.

Three of the eleven were size() - 1 underflows, rewritten as i + 1 < size():
ArcCircle's candidate scan (count is a size_t, so an empty point set gave
SIZE_MAX and a runaway loop rather than no iterations), and both deviation scans
in get_deviation_sum_squared. SVG::draw's thick-polyline loop got the same shape
for consistency even though its size >= 2 guard already made it safe.

The rest are type corrections. In Polyline.cpp the comparisons are against
buffer_size (a const int window parameter) and current_buffer_size, which the
surrounding asserts pin to arc.size() / line_length.size() / weights.size() -
both are genuine non-negative counts, so making the existing implicit conversion
explicit with size_t(...) changes nothing at runtime.

Verified. Per file: -fsyntax-only at the real build flags, sign-compare at 0, no
other flag's count moved. Runtime, against the binary from before any of this
campaign: bunny.stl is identical - 0 of 536 layers differ, max per-layer toolpath
delta 0.000000 mm, every role count and the whole summary equal. That holds
across all 10 pairs of 5 runs spanning both binaries.

3DBenchy is NOT evidence either way and is not being treated as such: 6
same-binary run pairs differ on 9 to 18 of 240 layers with a max delta of
3.346294 mm, and all six extrusion roles drift run-to-run. Baseline-vs-now lands
at 12 to 23 layers, the same distribution. The model cannot resolve a change this
small; bunny is the gate that decides, and bunny does not move.

+++ src/libslic3r/MultiPoint.cpp
@@ -511,7 +511,7 @@ Points MultiPoint::_douglas_peucker_plus(const Points& pts, const double toleran
             }
         }
     }
-    for(int i=1;i<result_pts.size();++i)
+    for(size_t i=1;i<result_pts.size();++i)
         assert(!result_pts[i - 1].coincides_with_epsilon(result_pts[i]));
     return result_pts;
 }
